refactor: 💡 move generate comp and generators serv to lib
migrate generate component component and generators service to
generators library, plus modified routing so at /generators/component we
will see  generate component component
