+++ libs/generators/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/generators/generators.component';
+export * from './lib/generators/generators/generate-component/generate-component.component';
